'move to folder' option for potentially-unused references
