+++ WolvenKit.App/Interaction/Interactions.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using WolvenKit.App.Helpers;
+using WolvenKit.App.Models.ProjectManagement.Project;
 using WolvenKit.App.ViewModels.Dialogs;
 
 namespace WolvenKit.App.Interaction;
@@ -47,12 +48,15 @@ public static class Interactions
     public static Func<string, string> Rename { get; set; } = _ => throw new NotImplementedException();
     public static Func<string, Tuple<string, bool>> RenameAndRefactor { get; set; } = _ => throw new NotImplementedException();
 
-    public static Func<(string, List<string>), List<string>> ShowDeleteFilesList { get; set; } = _ => throw new NotImplementedException();
+    public static Func<(string, List<string>, Cp77Project), (List<string>, string? moveToPath)> ShowDeleteOrMoveFilesList { get; set; } =
+        _ => throw new NotImplementedException();
 
     public static Func<(string, Dictionary<string, List<string>>), bool> ShowBrokenReferencesList { get; set; } =
         _ => throw new NotImplementedException();
 
     public static Func<string, string> AskForTextInput { get; set; } = _ => throw new NotImplementedException();
+
+    public static Func<(string, Cp77Project), string> AskForFolderPathInput { get; set; } = _ => throw new NotImplementedException();
 
     //custom views
     public static Func<bool> ShowFirstTimeSetup { get; set; } = () => throw new NotImplementedException();
