+++ WolvenKit.App/ViewModels/Shell/AppViewModel.cs
@@ -822,7 +822,7 @@ public partial class AppViewModel : ObservableObject/*, IAppViewModel*/
     [RelayCommand(CanExecute = nameof(CanShowProjectActions))]
     private async Task ScanForBrokenReferencePaths()
     {
-        _loggerService.Info($"Scanning {ActiveProject!.Files.Count} files. Please wait...");
+        _loggerService.Info($"Scanning {ActiveProject!.ModFiles.Count} files. Please wait...");
         _progressService.IsIndeterminate = true;
         var brokenReferences = await ActiveProject!.ScanForBrokenReferencePathsAsync(_archiveManager, _loggerService, _progressService);
 
@@ -895,28 +895,57 @@ public partial class AppViewModel : ObservableObject/*, IAppViewModel*/
         
     }
 
+    private async Task MoveFiles(List<string> filePaths, string destinationPath)
+    {
+        if (ActiveProject is null)
+        {
+            _loggerService.Warning("Can't do this without a project!");
+        }
+
+        foreach (var absolutePath in filePaths)
+        {
+            await ProjectResourceHelper.MoveAndRefactor(ActiveProject!, absolutePath, destinationPath, false);
+        }
+    }
+
     [RelayCommand(CanExecute = nameof(CanShowProjectActions))]
     private async Task FindUnusedFiles()
     {
-        _loggerService.Info($"Scanning {ActiveProject!.Files.Count(f => f.StartsWith("archive"))} files. Please wait...");
+        _loggerService.Info($"Scanning {ActiveProject!.ModFiles.Count} files. Please wait...");
 
         var allReferencePaths = await ActiveProject!.GetAllReferences(_progressService);
         var referencesHashSet = new HashSet<string>(allReferencePaths.SelectMany((r) => r.Value));
 
-        var unusedFiles =
-            ActiveProject!.ModFiles.Where(f => !referencesHashSet.Contains(ActiveProject.GetRelativePath(f))).ToList();
-        if (unusedFiles.Count == 0)
+        var potentiallyUnusedFiles = ActiveProject!.ModFiles
+            .Where(f => !referencesHashSet.Contains(ActiveProject.GetRelativePath(f))) // they're used
+            .Where(f => !_archiveManager.Lookup(f, ArchiveManagerScope.Basegame).HasValue) // they overwrite basegame files
+            .ToList();
+
+        if (potentiallyUnusedFiles.Count == 0)
         {
             _notificationService.ShowNotification("No un-used files in project", ENotificationType.Success, ENotificationCategory.App);
             return;
         }
 
+        // Filter out files by type
+
+        // potentiallyUnusedFiles.Where(f => _archiveManager.Lookup(f, ArchiveManagerScope.LocalProject).)
+        
+
         _progressService.Completed();
-        if (Interactions.ShowDeleteFilesList(("Delete un-used files?", unusedFiles)) is not { } deleteFiles || deleteFiles.Count == 0)
+        (var deleteFiles, var relativeDestPath) =
+            Interactions.ShowDeleteOrMoveFilesList(("Delete or move un-used files?", potentiallyUnusedFiles, ActiveProject));
+        if (deleteFiles.Count == 0)
         {
             return;
         }
 
+        if (relativeDestPath is not null)
+        {
+            await MoveFiles(deleteFiles, relativeDestPath);
+            return;
+        }
+        
         List<string> failedDeletions = [];
         foreach (var filePath in deleteFiles)
         {
