+++ WolvenKit.Core/Enums.cs
@@ -130,6 +130,8 @@ namespace WolvenKit.Common
     {
         Basegame,
         Mods,
+        BasegameAndMods,
+        LocalProject,
         Everywhere
     }
