+++ WolvenKit/Views/Dialogs/FolderPathInputDialogView.xaml
@@ -1,0 +1,58 @@
+<Window
+    x:Class="WolvenKit.Views.Dialogs.FolderPathInputDialogView"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:hc="https://handyorg.github.io/handycontrol"
+    xmlns:reactiveUi="http://reactiveui.net"
+    xmlns:syncfusion="http://schemas.syncfusion.com/wpf"
+    xmlns:helpers="clr-namespace:WolvenKit.Functionality.Helpers"
+    Title="{Binding Title, RelativeSource={RelativeSource Self}}"
+    SizeToContent="WidthAndHeight"
+    FocusManager.FocusedElement="{Binding ElementName=TextBox}"
+    WindowStartupLocation="CenterScreen">
+
+    <Grid
+        Margin="5,5,2,5"
+        HorizontalAlignment="Stretch"
+        VerticalAlignment="Stretch"
+        hc:ThemeManager.RequestedAccentColor="{DynamicResource MahApps.Brushes.Accent3}">
+
+        <Grid.Resources>
+            <ResourceDictionary>
+                <ResourceDictionary.MergedDictionaries>
+                    <hc:ThemeResources />
+                    <hc:Theme />
+                </ResourceDictionary.MergedDictionaries>
+            </ResourceDictionary>
+        </Grid.Resources>
+
+
+        <syncfusion:WizardControl CancelButtonCancelsWindow="True" FinishButtonClosesWindow="True">
+            <syncfusion:WizardPage
+                BackVisibility="Collapsed"
+                CancelVisibility="Visible"
+                FinishVisibility="Visible"
+                HelpVisibility="Collapsed"
+                NextVisibility="Collapsed"
+                PreviewKeyDown="WizardPage_PreviewKeyDown"
+                PageType="Exterior">
+                <Grid
+                    HorizontalAlignment="Stretch"
+                    VerticalAlignment="Stretch">
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition Width="600" />
+                    </Grid.ColumnDefinitions>
+                    <Grid.RowDefinitions>
+                        <RowDefinition Height="25" />
+                    </Grid.RowDefinitions>
+
+                    <TextBox x:Name="TextBox" KeyUp="Textbox_OnKeyUp" Style="{StaticResource ValidatingControlStyle}" />
+
+                </Grid>
+            </syncfusion:WizardPage>
+        </syncfusion:WizardControl>
+
+    </Grid>
+
+
+</Window>
