+++ WolvenKit/Views/Dialogs/Windows/DeleteOrMoveFilesListDialogView.xaml
@@ -1,5 +1,5 @@
 ﻿<Window
-    x:Class="WolvenKit.Views.Dialogs.Windows.DeleteFilesListDialogView"
+    x:Class="WolvenKit.Views.Dialogs.Windows.DeleteOrMoveFilesListDialogView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:hc="https://handyorg.github.io/handycontrol"
@@ -43,14 +43,16 @@
             CancelButtonCancelsWindow="True"
             FinishButtonClosesWindow="True"
             CancelText="Close"
-            FinishText="Delete unused files">
+            FinishText="Delete unused files"
+            NextText="Move to folder"
+            NextEnabled="{Binding HasFiles}"
+            Next="OnNextButtonClicked">
             <syncfusion:WizardPage
                 BackVisibility="Collapsed"
                 CancelVisibility="Visible"
                 FinishVisibility="Visible"
                 HelpVisibility="Collapsed"
-                NextVisibility="Collapsed"
-
+                NextVisibility="Visible"
                 PreviewKeyDown="WizardPage_PreviewKeyDown"
                 PageType="Exterior">
                 <Grid
@@ -77,7 +79,9 @@
                             TextWrapping="Wrap"
                             VerticalScrollBarVisibility="Auto"
                             VerticalAlignment="Top"
-                            Text="{Binding Files, Converter={StaticResource ListToStringConverter}}" />
+                            VerticalContentAlignment="Top"
+                            KeyUp="DeleteFilesTextbox_OnKeyUp"
+                            Text="{Binding Files, Converter={StaticResource ListToStringConverter}, Mode=TwoWay}" />
                     </StackPanel>
 
                     <StackPanel Grid.Row="2" Orientation="Vertical">
